Time: 0 ms (100%), Space: 9.1 MB (42.8%) - LeetHub

+++ 0067-add-binary/0067-add-binary.cpp
@@ -26,12 +26,12 @@ public:
     string addBinary(string a, string b) {
         
         
-        int carry = 0;
+        short carry = 0;
         if (!(a.size() > b.size()))
             swap(a, b);
         string result = a;
-        int i = a.size() - 1;
-        int j = b.size() - 1;
+        short i = a.size() - 1;
+        short j = b.size() - 1;
         short sum;
         //    "100"              100
         // "110010"           110010
